Move chapter 2 exercises and solutions into standard layout.
