+++ src/test/kotlin/chapter2/exercises/ex1/listing.kt
@@ -1,22 +1,19 @@
-package chapter2.solutions
+package chapter2.exercises.ex1
 
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 import kotlinx.collections.immutable.persistentMapOf
 
-class Solution_2_1 : WordSpec({
+class Exercise1 : WordSpec({
     // tag::init[]
-    fun fib(i: Int): Int {
-        tailrec fun go(cnt: Int, curr: Int, nxt: Int): Int =
-            if (cnt == 0)
-                curr
-            else go(cnt - 1, nxt, curr + nxt)
-        return go(i, 0, 1)
-    }
+    fun fib(i: Int): Int = TODO()
     // end::init[]
 
+    /**
+     * Re-enable the tests by removing the `!` prefix!
+     */
     "fib" should {
-        "return the nth fibonacci number" {
+        "!return the nth fibonacci number" {
             persistentMapOf(
                 Pair(1, 1),
                 Pair(2, 1),
