+++ src/test/kotlin/chapter2/exercises/ex3/listing.kt
@@ -1,18 +1,19 @@
-package chapter2.solutions
+package chapter2.exercises.ex3
 
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 
-class Solution_2_3 : WordSpec({
+class Exercise3 : WordSpec({
     // tag::init[]
-    fun <A, B, C> curry(f: (A, B) -> C): (A) -> (B) -> C =
-        { a: A -> { b: B -> f(a, b) } }
+    fun <A, B, C> curry(f: (A, B) -> C): (A) -> (B) -> C = TODO()
     // end::init[]
 
+    /**
+     * Re-enable the tests by removing the `!` prefix!
+     */
     "curry" should {
-        """break down a function that takes multiple arguments
-            into a series of functions that each take only
-            one argument""" {
+        """!break down a function that takes multiple arguments into
+            a series of functions that each take only oneargument""" {
 
                 val f: (Int) -> (Int) -> String =
                     curry { a: Int, b: Int -> "$a:$b" }
