+++ src/test/kotlin/chapter2/exercises/ex5/listing.kt
@@ -1,16 +1,18 @@
-package chapter2.solutions
+package chapter2.exercises.ex5
 
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 
-class Solution_2_5 : WordSpec({
+class Exercise5 : WordSpec({
     // tag::init[]
-    fun <A, B, C> compose(f: (B) -> C, g: (A) -> B): (A) -> C =
-        { a: A -> f(g(a)) }
+    fun <A, B, C> compose(f: (B) -> C, g: (A) -> B): (A) -> C = TODO()
     // end::init[]
 
+    /**
+     * Re-enable the tests by removing the `!` prefix!
+     */
     "compose" should {
-        "apply function composition over two functions" {
+        "!apply function composition over two functions" {
             val fahrenheit2celsius: (Double) -> String =
                 compose<Double, Double, String>(
                     { b -> "$b degrees celsius" },
