+++ src/test/kotlin/chapter2/solutions/sol1/listing.kt
@@ -1,19 +1,22 @@
-package chapter2.exercises
+package chapter2.solutions.sol1
 
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 import kotlinx.collections.immutable.persistentMapOf
 
-class Exercise_2_1 : WordSpec({
+class Solution1 : WordSpec({
     // tag::init[]
-    fun fib(i: Int): Int = TODO()
+    fun fib(i: Int): Int {
+        tailrec fun go(cnt: Int, curr: Int, nxt: Int): Int =
+            if (cnt == 0)
+                curr
+            else go(cnt - 1, nxt, curr + nxt)
+        return go(i, 0, 1)
+    }
     // end::init[]
 
-    /**
-     * Re-enable the tests by removing the `!` prefix!
-     */
     "fib" should {
-        "!return the nth fibonacci number" {
+        "return the nth fibonacci number" {
             persistentMapOf(
                 Pair(1, 1),
                 Pair(2, 1),
