+++ src/test/kotlin/chapter2/solutions/sol5/listing.kt
@@ -1,18 +1,16 @@
-package chapter2.exercises
+package chapter2.solutions.sol5
 
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.WordSpec
 
-class Exercise_2_5 : WordSpec({
+class Solution5 : WordSpec({
     // tag::init[]
-    fun <A, B, C> compose(f: (B) -> C, g: (A) -> B): (A) -> C = TODO()
+    fun <A, B, C> compose(f: (B) -> C, g: (A) -> B): (A) -> C =
+        { a: A -> f(g(a)) }
     // end::init[]
 
-    /**
-     * Re-enable the tests by removing the `!` prefix!
-     */
     "compose" should {
-        "!apply function composition over two functions" {
+        "apply function composition over two functions" {
             val fahrenheit2celsius: (Double) -> String =
                 compose<Double, Double, String>(
                     { b -> "$b degrees celsius" },
